Commit check now behaves like IslaVista.
Note that commit check only applies in the following cases:

1) On nodes or ancestors of nodes that have been set or modified
2) Does not applied to deleted nodes

Feature developers should:

1) locate commit checks at a common ancestor if checking across a hierarchy
2) expect that the commit check only enforces created behavior (therefore behaves more like a create check).

An additional flag is provided "-f" to allow for full syntax and commit checks. This allows a complete single pass on all syntax and
commit errors. The default behavior is to stop processing on the first error.