The SYNTAX and COMMIT tags will be executed before an actions (UPDATE, CREATE, DELETE, END, BEGIN) within a priority tree.
That means that all work (include the commit-check pass) will be
performed on a priority tree before proceeding to the next priority tree.

If a priority tree generates a failure in the commit-check pass
processing will halt on the priority tree and no further processing will
be performed on the priority tree. The failed priority tree will NOT be
copied to the active working directory.

All nodes in the priority tree will execute SYNTAX and COMMIT actions
except for nodes that have been deleted. Note this iteration is
performed on the working tree, so all commit-check nodes will either be
directly modified or have a node down the hierarchy that had been modified.

The working directory and active directory will not be modified until
after all priority trees have been processed.