Bugfix 4700, 4269: Fix set and commit-time checks of ethernet address values
Changed the the set-time and commit-time check of ethernet interface address
values.  These checks need to prevent configuring both DHCP and static
IPv4 addresses on the same interfac.  The previous checks were comparing
against the running configuration tree, not the proposed config tree.  Now
the set-time check is purely a syntax check, and the commit-time check
only checks for both DHCP and static IPv4 addresses in the proposed config.
The system now allows DHCP and static IPv6 addresses to be configured on
the same interface.