Change how floppy is mounted
This fixes several problems:
 1) Only try mounting once which avoids several I/O errors
 2) Mount floppy with safer options for security
 3) Use more optimum methods to reduce boot time

This should really be done by better udev support. But the root of the
problem lies in the Debian upstream package maintainer.
He is hostile to later advancements, and still maintains a static
/dev infrastructure as well as devfs compatiablity!
Ubuntu, SUSE, and Redhat have all migrated to creating devices via udev;
this would allow floppy device to only exist if floppy is present.
And then inserting a floppy could cause udev add event to mount
/media/floppy...