Workaround 5100: Disable TSO on NICs using the igb driver.
Packets forwarded by NICs using the igb driver are corrupted when
TSO is enabled.  It appears that a train of back-to-back packets from the
same flow coalesced by LRO are not correctly re-generated by TSO.  This
workaround prevents the problem by disabling TSO.

This problem does not occur in the Kenwood code base, so this workaround
should NOT be merged forward to Kenwood.