Multiple (i.e. one or two) datatypes can now be supported on a node.def type definition.
type: ipv4,ipv6

Will now allow for the value to validated against both data types (or the first success). In the case of both failing the following message format will be returned:

vyatta@vyatta# set load-balancing wan rule 1
"1" is not a valid value of type "bool"
"1" is not a valid value of type "ipv4"

Note, that any combination of types with text being one of the types is not supported. The reason being is that with text any value is allowed, therefore it doesn't really
make sense to validate a datatype on text + ipv4 when text will always be successful.