support for PARENT reference in priority statements. Example:
priority: PARENT

Means that this priority group will adopt the priority value of the parent. And the behavior on deletion of this priority root node will be deleted before the parent. On
all other operations this node will be scheduled after the parent.

This should fix problems in scheduling creation/deletion orders for priority groups so long as they are located within the same hierarchy.

priority.pl has been updated as well.