Bugfix 5830: Teach auto affinity script to use new queue naming scheme.
The latest version of the ixgbe driver uses a scheme whereby each pair
of TX and RX queues is assigned to the same IRQ.  Previously, the TX
and RX queues could be assigned to IRQs independently. In
/proc/interrupts, each pair of queues is named ethX-TxRx-Y, where X is
the ethernet interface unit number, and Y is the TX and RX queue
number.