vyatta-cfg: display openvpn interface descriptions after they reset
When OpenVPN tunnels are taken down and recreated by the OpenVPN
process, any previously set descriptions on the interfaces are missing.
This is caused by the fact that the ifalias for that interface isn't
readded when the interface is recreated.

This patch updates the interface_description function so that it either
returns the current value of ifalias if it's set on the interface, or
uses the value set in the active config if it's present.

Bug #7 http://bugzilla.vyos.net/show_bug.cgi?id=7
Bug #313 http://bugzilla.vyos.net/show_bug.cgi?id=313