vyatta-cfg: fix for "failed to generate commited config"
When changes are made to the system by more than one user, only the
first user after a reboot is able to commit successfully, the others
fail with "Failed to generate committed config".  This is caused as
a result of the underlying unionfs-fuse mount used for the system
configuration and the fact that a .unionfs directory if left in the
active configuration with permissions set such that only the original
committing user can change them.

Bug #403 http://bugzilla.vyos.net/show_bug.cgi?id=403