vyatta-cfg: remove unused unionfs mounts following unclean config exits
If a configuration session is exited uncleanly, either as a result of a
network disconnect / timeout or other forced quit from within config
mode, the unionfs mount is left in place.  These remain until the
system is rebooted, giving an error when unmounting the config
directory.

This commit checks for old configuartion directories when a user enters
config mode, removing any found that are owned by that user but not
part of any active configuration seesion.

Bug #564 http://bugzilla.vyos.net/show_bug.cgi?id=564