Add helper function is_dhcp_4or6_enabled to replace is_dhcp_enabled
The function `is_dhcp_enabled` ignores DHCPv6. It is likely that many of
the uses of this function are not functioning properly, or functioning
only by luck, for DHCPv6 users. As DHCPv6 usage increases, the problems
will become more evident. This new function, `is_dhcp_4or6_enabled`,
should be safely and gradually migrated, checking that each replacement
does not break any functionality.

Additionally, `is_address_enabled` is incorrectly ignoring DHCPv6 as
well. I believe the risk to changing this function to be minimal,
as I could not find any actual uses of it. This change will safegaurd
future uses.