T1899: fix for unionfs-fuse rename of meta directory
In release v1.0 of unionfs-fuse, the meta directory was renamed from
.unionfs to .unionfs-fuse. Consequently, the clean up of the directory
following a commit vacuously succeeds, and the meta data remains. On a
subsequent commit, this hidden directory will be copied into the tmp
directory structure. As permissions on the whiteout files belong solely
to the user, commit actions in a multiuser environment may encounter
permission errors, leading to a failure in commit. Change the meta
directory name here to accord with the change in unionfs-fuse, but
beware that the change is reverted in release v2.0 of unionfs-fuse.