Cloud-Init: T2117: Added order dependency for cloud-config.service
As part of cleaning up our Cloud-Init package from changes incompatible with the upstream package, the `Before=vyos-router.service` should be moved from the `cloud-config.service` file to the `vyos-router.service` as `After=cloud-config.service`.
This should not affect anyhow the `vyos-router.service`, it will work as before in systems without Cloud-Init.
In systems with Cloud-Init package the `vyos-router.service` should be run after the `cloud-config.service`.