scripts: T3425: Make paths to custom scripts relative
In some cases, like ISO or PXE boot, the `/opt/vyatta/etc/config/` folder does not bind to `/config/`. To run scripts from the `/scripts/` subfolder paths must be relative, just like for the `config.boot` file.
This commit replaces hardcoded paths to relative that use the `$vyatta_sysconfdir` as the prefix.