cli: T6769: fix after commit args validation
Command `commit-config` passed all it's arguments
to `commit`. This led to an error in case of simple usage
`commit-config 1`: "1" was passed to commit and validation
rejected this argument.

Fix is just to use `args` variable instead that was already
calculated in function and filtered out first numerical argument,
but wasn't used.